+++ requirements.txt
@@ -7,3 +7,4 @@ scikit_learn==1.0
 keras==2.6.0
 tensorflow==2.6.0
 xgboost==1.4.2
+sbrscrape==0.0.2
